move getFingerState
